LSM/ManifestLog: Append to superblock manifest on block close
When there were multiple manifest logs, this needed to be done in `flush()` (synchronous with respect to the start of the bar) to avoid determinism issues as different trees raced.
But with a single manifest log, this ordering is guaranteed regardless – the append is more naturally part of `close_block()`.